LineOffset Fix

+++ src/main/java/com/subu/fileprocessor/dao/Offset.java
@@ -3,11 +3,13 @@ package com.subu.fileprocessor.dao;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigInteger;
+
 
 @Data
 @RequiredArgsConstructor
 public class Offset {
-    private final int lineOffset;
-    private final int charOffset;
+    private final BigInteger lineOffset;
+    private final BigInteger charOffset;
 }
